UPSTREAM: drm/amd/display: Keep GSL for full updates with planes that flip VSYNC
[Why]
When enabling PIP in Heaven, the PIP planes are VSYNC
flip and is also the top-most pipe. In this case GSL
will be disabled because we only check immediate flip
for the top pipe. However, the desktop planes are still
flip immediate so we should at least keep GSL on until
the full update.

[How]
Check each pipe in the tree to see if any planes
are flip immediate. Maintain the GSL lock if yes,
and take it down after when unlocking if any planes
are flipping VSYNC. Keeping GSL on with VSYNC +
flip immediate planes causes corruption.

Signed-off-by: Alvin Lee <alvin.lee2@amd.com>
Reviewed-by: Aric Cyr <Aric.Cyr@amd.com>
Acked-by: Qingqing Zhuo <qingqing.zhuo@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
(cherry picked from commit 6f2239ccdfc04938dc35e67dd60191b2c05dfb63)
Signed-off-by: Sean Paul <seanpaul@chromium.org>

BUG=b:177892854
TEST=Tested on volteer, build-tested trogdor

Change-Id: I394a1bc4e952fa0b045665bf5a9045e041f3fa3d