CHROMIUM: config: Enable CONFIG_USB_NET_AQC111
This is a driver for several USB to 5GBase-T ethernet adapters
in the wild.

Previous kernels used cdc_ether for these devices, which was already
enabled, so it is not necessary to backport this to older kernels.
For v5.4 in particular, this is required because changes in the cdc_ether
and this driver meant that the QNAP device and others are blocklisted from
cdc_ether as the new acq11 driver is intended to handle AQC111 chipset
devices from now on.

Signed-off-by: Benson Leung <bleung@chromium.org>

BUG=b:159662671
TEST=Attach QNAP QNA-UC5G1T USB to 5GbE Adapter to Volteer's USB-C port.
Check that it is recognized in the kernel, and shows up as a valid network
interface.

Originally-Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2518145
Reviewed-by: Grant Grundler <grundler@chromium.org>
Reviewed-by: Alex Levin <levinale@google.com>
Commit-Queue: Benson Leung <bleung@google.com>
Tested-by: Benson Leung <bleung@google.com>
Auto-Submit: Benson Leung <bleung@chromium.org>
(cherry picked from commit b40547425bb386a831028bf41e189f3c0ab6b736)

Change-Id: If93530d56033c1f928174061e5513c4af6144766
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2519110
Reviewed-by: Benson Leung <bleung@google.com>
Reviewed-by: Benson Leung <bleung@chromium.org>
Commit-Queue: Alex Levin <levinale@google.com>
Tested-by: Alex Levin <levinale@google.com>
Tested-by: Benson Leung <bleung@chromium.org>
Auto-Submit: Alex Levin <levinale@google.com>