drm/msm: dsi: Constify dsi_host_ops
The only usage of dsi_host_ops is to assign its address to the ops field
in the mipi_dsi_host struct, which is a const pointer. Make it const to
allow the compiler to put it in read-only memory.

Signed-off-by: Rikard Falkeborn <rikard.falkeborn@gmail.com>
Signed-off-by: Rob Clark <robdclark@chromium.org>
(cherry picked from commit 8b6947a81e56f0ef06be8a6d805df9634b4fd3f7)
Signed-off-by: Sean Paul <seanpaul@chromium.org>