FROMGIT: drm: panel: simple: Defer unprepare delay till next prepare to shorten it
It is believed that all of the current users of the "unprepare" delay
don't actually need to wait the amount of time specified directly in
the unprepare phase.  The purpose of the delay that's specified is to
allow the panel to fully power off so that we don't try to power it
back on before it's managed to full power down.

Let's use this observation to avoid the fixed delay that we currently
have.  Instead of delaying, we'll note the current time when the
unprepare happens.  If someone then tries to prepare the panel later
and not enough time has passed, we'll do the delay before starting the
prepare phase.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Sam Ravnborg <sam@ravnborg.org>
Link: https://patchwork.freedesktop.org/patch/msgid/20201109170018.v4.2.I06a95d83e7fa1bd919c8edd63dacacb5436e495a@changeid
(cherry picked from commit e5e30dfcf3db1534019c40d94ed58fd2869f9359
 https://github.com/freedesktop/drm-misc.git drm-misc-next)

BUG=b:169288627
TEST=Panel still works

Change-Id: I06a95d83e7fa1bd919c8edd63dacacb5436e495a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2577664
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Rob Clark <robdclark@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>
(cherry picked from commit ef5445288e6c12bbb672e102a1c73fcb0dd9e961)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2716960
Reviewed-by: Douglas Anderson <dianders@chromium.org>